# asset/swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: 2023 asset management
  description: 2023 asset management
  contact:
    name: Kevin Chen
    email: kevinyay945@gmail.com
    url: https://kevinyay945.com
servers:
  - url: http://localhost:18191
tags:
  - name: v1Asset
    description: version 1 Asset
paths:
  /v1/temp-link/{location}/{fileName}:
    get:
      summary: redirect to the public link
      description: redirect to the public link
      tags:
        - v1Asset
      operationId: v1RedirectToPublicLink
      parameters:
        - name: location
          in: path
          description: save location
          required: true
          schema:
            $ref: '#/components/schemas/V1AssetLocation'
        - name: fileName
          in: path
          description: save location
          required: true
          schema:
            type: string
      responses:
        '302':
          description: public link
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      
  /v1/asset/{location}:
    post:
      summary: upload one asset and get preview link
      description: upload one asset and get preview link
      tags:
        - v1Asset
      operationId: v1UploadAsset
      parameters:
        - name: location
          in: path
          description: save location
          required: true
          schema:
            $ref: '#/components/schemas/V1AssetLocation'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: license data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TempAsset'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    V1AssetLocation:
      type: string
      enum:
        - "obsidian"
        - "blog"
      x-enum-varnames:
        - V1AssetLocationObsidian
        - V1AssetLocationBlog
    TempAsset:
      required:
        - url
      properties:
        url:
          type: string
          example: "http://localhost/temp-link"
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message