# interface/rest_api/openapi_api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// upload one asset and get preview link
	// (POST /v1/asset/{location})
	V1UploadAsset(ctx echo.Context, location V1AssetLocation) error
	// redirect to the public link
	// (GET /v1/temp-link/{location}/{fileName})
	V1RedirectToPublicLink(ctx echo.Context, location V1AssetLocation, fileName string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// V1UploadAsset converts echo context to params.
func (w *ServerInterfaceWrapper) V1UploadAsset(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "location" -------------
	var location V1AssetLocation

	err = runtime.BindStyledParameterWithLocation("simple", false, "location", runtime.ParamLocationPath, ctx.Param("location"), &location)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter location: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.V1UploadAsset(ctx, location)
	return err
}

// V1RedirectToPublicLink converts echo context to params.
func (w *ServerInterfaceWrapper) V1RedirectToPublicLink(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "location" -------------
	var location V1AssetLocation

	err = runtime.BindStyledParameterWithLocation("simple", false, "location", runtime.ParamLocationPath, ctx.Param("location"), &location)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter location: %s", err))
	}

	// ------------- Path parameter "fileName" -------------
	var fileName string

	err = runtime.BindStyledParameterWithLocation("simple", false, "fileName", runtime.ParamLocationPath, ctx.Param("fileName"), &fileName)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter fileName: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.V1RedirectToPublicLink(ctx, location, fileName)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/v1/asset/:location", wrapper.V1UploadAsset)
	router.GET(baseURL+"/v1/temp-link/:location/:fileName", wrapper.V1RedirectToPublicLink)

}