# interface/rest_api/openapi_types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// Defines values for V1AssetLocation.
const (
	V1AssetLocationBlog     V1AssetLocation = "blog"
	V1AssetLocationObsidian V1AssetLocation = "obsidian"
)

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// TempAsset defines model for TempAsset.
type TempAsset struct {
	Url string `json:"url"`
}

// V1AssetLocation defines model for V1AssetLocation.
type V1AssetLocation string

// V1UploadAssetMultipartBody defines parameters for V1UploadAsset.
type V1UploadAssetMultipartBody struct {
	Image *openapi_types.File `json:"image,omitempty"`
}

// V1UploadAssetMultipartRequestBody defines body for V1UploadAsset for multipart/form-data ContentType.
type V1UploadAssetMultipartRequestBody V1UploadAssetMultipartBody